HLRC: Split up RequestConverters
The RequestConverters were becoming quite large, so this splits them
up. Besides creating the named classes, this was a mechanical move in
the IDE.
